+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased
 
 ### Bug Fixes / Nits
+- Fix elasticsearch hybrid scoring (#7852)
 - Replace `get_color_mapping` and `print_text` Langchain dependency with internal implementation (#7845)
+
 
 ## [0.8.35] - 2023-09-27